Chairs create the sections, papers to sections (getSections)

+++ catalog-core/src/main/java/ro/ubb/catalog/core/service/PCMemberService.java
@@ -216,7 +216,7 @@ public class PCMemberService {
             return listeners.stream().filter(l -> l.getSectionId().equals(s)).count();
         }
 
-        return (long) (0);
+        return 0;
     }
 }
